refactor: move Combined to examples dir
